Avoid using std::chono::file_clock
There's a big gap in GCC - there's file_clock in the std::filesystem,
but there's no way to convert that clock to the system_clock and
use to_time_t(). This avoids usage by issuing the stat directly.

Note that the similar will have to be done in Windows. We have to
convert to the single epoch to be able to compare the files
cross-platform.